Modified the verify, change password, and reset password controllers

+++ controllers/userController.js
@@ -6,6 +6,8 @@ const CustomError = require('../models/customError');
 const { isValidObjectId } = require('mongoose');
 const { getStorage, ref, uploadBytesResumable, getDownloadURL } = require('firebase/storage');
 const app = require('../services/firebase');
+const validator = require('validator');
+const bcrypt = require('bcrypt');
 
 async function emailSent(otp, email) {
     return await sendEmail(email, 'Verify your email address', `Enter the following code in the application to verify your account:\n 
@@ -87,7 +89,7 @@ const register = async (req, res) => {
 
             if (!sent) throw new CustomError("Unable to send email verification otp", 500);
 
-            return jsonResponse(res, { message: 'Otp sent to your email please verify account' }, 201)
+            return jsonResponse(res, { message: 'Otp sent to your email please verify account', user_id: user._id }, 201)
         }
     } catch (error) {
         // console.log(error.message)
@@ -97,12 +99,13 @@ const register = async (req, res) => {
 
 const verify = async (req, res) => {
     try {
-        const { email, token } = req.params;
+        const { uid, token } = req.params;
         const type = req.query.type;
+        if (!isValidObjectId(uid)) throw new CustomError('Not a valid user id', 400);
 
         if (type === 'password-reset-verify') {
 
-            const user = await User.findOne({ email });
+            const user = await User.findOne({ _id: uid });
             if (!user) throw new CustomError('User not found', 404);
 
             const passwordToken = await Token.findOne({ user_id: user._id, token, for_pass: true });
@@ -112,13 +115,13 @@ const verify = async (req, res) => {
 
         } else if (type === 'email-verify') {
 
-            const user = await User.findOne({ email });
+            const user = await User.findOne({ _id: uid });
             if (!user) throw new CustomError('Invalid OTP', 401);
 
             const emailVerifyToken = await Token.findOne({ user_id: user._id, token, for_pass: false });
             if (!emailVerifyToken) throw new CustomError('Invalid OTP', 400);
 
-            await User.updateOne({ email }, { $set: { verified: true } });
+            await User.updateOne({ _id: user._id }, { $set: { verified: true } });
             await Token.deleteOne({ user_id: user._id, for_pass: false });
 
             jsonResponse(res, { message: "Email verified successfully" });
@@ -132,20 +135,24 @@ const verify = async (req, res) => {
 
 const changePassword = async (req, res) => {
     try {
-        const { email, token } = req.params;
+        const { uid, token } = req.params;
         const { password } = req.body;
+        if (!validator.isStrongPassword(password)) throw new CustomError('Password must contain a number, a special character, a capital letter, and must be 8 characters long', 400);
+        if (!isValidObjectId(uid)) throw new CustomError('Invalid user id', 401);
 
-        const user = User.findOne({ email });
+        const user = await User.findOne({ _id: uid });
+        if (!user) throw new CustomError('Unauthorized request', 401);
 
         const passwordToken = await Token.findOne({ user_id: user._id, token });
+
         if (!passwordToken) throw new CustomError('OTP has expired', 400);
 
         const hash = await bcrypt.hash(password, Number(process.env.SALT));
 
         if (user.verified) {
-            await User.updateOne({ email }, { $set: { password: hash } });
+            await User.updateOne({ _id: user._id }, { $set: { password: hash } });
         } else {
-            await User.updateOne({ email }, { $set: { password: hash, verified: true } });
+            await User.updateOne({ _id: user._id }, { $set: { password: hash, verified: true } });
         }
 
         await Token.deleteOne({ user_id: user._id });
@@ -159,24 +166,26 @@ const changePassword = async (req, res) => {
 const resetPassword = async (req, res) => {
     try {
         const { email } = req.body;
+        if (!validator.isEmail(email)) throw new CustomError('Enter a valid email address', 400);
 
         const user = await User.findOne({ email });
         if (!user) throw new CustomError('Email not found', 404);
 
-        var token = await Token.findOne({ user_id: user._id });
+        const available = await Token.findOne({ user_id: user._id });
+        if (available) await Token.deleteOne({ user_id: user._id });
 
-        if (token) await Token.deleteOne({ user_id: user._id });
+        const token = await Token.create({ user_id: user._id, token: createOtp(), for_pass: true });
 
-        token = await Token.create({ email, token: createOtp() });
+        const sent = await emailSent(token.token, email);
 
-        if (await emailSent(token.token, email)) {
-            jsonResponse(res, { message: "Otp sent to your email" })
+        if (sent) {
+            jsonResponse(res, { message: "Otp sent to your email", user_id: user._id })
         } else {
             throw new CustomError('Unable to send otp to your email');
         }
 
     } catch (error) {
-        // console.log(error)
+        console.log(error)
         errorResponse(res, error);
     }
 }
